[Decomp][Player] Cohesion split: extract spell-cooldown methods to PlayerCooldown.cpp
